Create Userbase-Queries
This project involves Netflix Userbase (source: keggle). Ten specific information was scraped on the userbase by generating SQL queries. 

This project was done on MySQL using the TablePlus client. Web-based hosting was used due to the limitations on the devide used (this project was done on an iPad Pro M1).