Add files via upload
When the data was sourced, the CSV file needed to be cleaned up before importing to MySQL. This action was done using Google Sheets. Due to that, you may encounter Error Values when opening it up on Microsoft Excel. However, you can access the main sheets document on the link below. 

https://docs.google.com/spreadsheets/d/1Va6HVX6mH2V0ofRIWWIfWXf0C8EWyiCkzdUzBABbCOA/edit?usp=sharing